+++ MyClaw/Objects/BearSailor.h
@@ -10,7 +10,7 @@ public:
 
 	void Logic(uint32_t elapsedTime) override;
 	Rectangle2D GetRect() override;
-	pair<Rectangle2D, int8_t> GetAttackRect() override;
+	pair<Rectangle2D, int> GetAttackRect() override;
 
 private:
 	void makeAttack() override;
