+++ MyClaw/Objects/Rat.h
@@ -10,7 +10,7 @@ public:
 	Rat(const WwdObject& obj, Player* player);
 
 	void makeAttack() override;
-	pair<Rectangle2D, int8_t> GetAttackRect() override;
+	pair<Rectangle2D, int> GetAttackRect() override;
 
 	bool isDuck() const override;
 	bool isTakeDamage() const override;
@@ -27,7 +27,7 @@ public:
 	void Logic(uint32_t elapsedTime) override;
 	void Draw() override;
 	Rectangle2D GetRect() override;
-	pair<Rectangle2D, int8_t> GetAttackRect() override;
+	pair<Rectangle2D, int> GetAttackRect() override;
 
 private:
 	Cannon* _cannon;
