+++ MyClaw/Objects/Siren.h
@@ -9,7 +9,7 @@ public:
 	Siren(const WwdObject& obj, Player* player);
 
 	Rectangle2D GetRect() override;
-	pair<Rectangle2D, int8_t> GetAttackRect() override;
+	pair<Rectangle2D, int> GetAttackRect() override;
 
 	void makeAttack() override;
 };
